StlIO: moved from math/ to utils/
as it's a set of STL serialization utils
